fix(fmp): remove redundant u16 cast caught by clippy
payload_len is already u16 at line 154, so the second as u16 cast
on line 170 is unnecessary. Clippy -D warnings caught this.

+++ crates/microfips-core/src/fmp.rs
@@ -167,7 +167,7 @@ pub fn build_established(
         return None;
     }
 
-    let prefix = build_prefix(PHASE_ESTABLISHED, 0x00, payload_len as u16);
+    let prefix = build_prefix(PHASE_ESTABLISHED, 0x00, payload_len);
     out[..COMMON_PREFIX_SIZE].copy_from_slice(&prefix);
     let mut pos = COMMON_PREFIX_SIZE;
